feat!: rename sort-map-elements rule to sort-maps
