+++ docs/rules/index.md
@@ -18,7 +18,7 @@ description: ESLint Plugin Perfectionist list of rules
 | [sort-imports](/rules/sort-imports)               | enforce sorted imports                      | 🔧 |
 | [sort-interfaces](/rules/sort-interfaces)         | enforce sorted interface properties         | 🔧 |
 | [sort-jsx-props](/rules/sort-jsx-props)           | enforce sorted JSX props                    | 🔧 |
-| [sort-map-elements](/rules/sort-map-elements)     | enforce sorted Map elements                 | 🔧 |
+| [sort-maps](/rules/sort-maps)                     | enforce sorted Map elements                 | 🔧 |
 | [sort-named-exports](/rules/sort-named-exports)   | enforce sorted named exports                | 🔧 |
 | [sort-named-imports](/rules/sort-named-imports)   | enforce sorted named imports                | 🔧 |
 | [sort-object-types](/rules/sort-object-types)     | enforce sorted object types                 | 🔧 |
